+++ mfront/src/BehaviourDSLCommon.cxx
@@ -4899,7 +4899,8 @@ namespace mfront{
        << "#include\"TFEL/Material/MechanicalBehaviourTraits.hxx\"\n"
        << "#include\"TFEL/Material/OutOfBoundsPolicy.hxx\"\n"
        << "#include\"TFEL/Material/BoundsCheck.hxx\"\n"
-       << "#include\"TFEL/Material/Lame.hxx\"\n";
+       << "#include\"TFEL/Material/Lame.hxx\"\n"
+       << "#include\"TFEL/Material/Hosford.hxx\"\n";
     if(this->mb.getSymmetryType()==ORTHOTROPIC){
       os << "#include\"TFEL/Material/Hill.hxx\"\n"
 	 << "#include\"TFEL/Material/OrthotropicAxesConvention.hxx\"\n";