+++ docs/web/news.md
@@ -15,6 +15,56 @@
 \newcommand{\sigmaH}{\sigma_{H}}
 \newcommand{\Frac}[2]{{{\displaystyle \frac{\displaystyle #1}{\displaystyle #2}}}}
 \newcommand{\deriv}[2]{{\displaystyle \frac{\displaystyle \partial #1}{\displaystyle \partial #2}}}
+\newcommand{\sigmaeq}{\sigma_{\mathrm{eq}}}
+
+# Hosford equivalent stress (17/10/2017)
+
+The header `TFEL/Material/Hosford.hxx` introduces three functions
+which are meant to compute the Hosford equivalent stress and its first
+and second derivatives. *This header is automatically included by
+`MFront`*
+
+The Hosford equivalent stress is defined by:
+\[
+\sigmaeq=\sqrt[a]{\Frac{1}{2}\paren{\paren{s_{1}-s_{2}}^{a}+\paren{s_{1}-s_{3}}^{a}+\paren{s_{2}-s_{3}}^{a}}}
+\]
+where \(s_{1}\), \(s_{2}\) and \(s_{3}\) are the eigenvalues of the
+stress.
+
+Therefore, when \(a\) goes to infinity, the Hosford stress reduces to
+the Tresca stress. When \(n = 2\) the Hosford stress reduces to the
+von Mises stress.
+
+The following function has been implemented:
+
+- `computeHosfordStress`: return the Hosford equivalent stress
+- `computeHosfordStressNormal`: return a tuple containg the Hosford
+  equivalent stress and its first derivative (the normal)
+- `computeHosfordStressSecondDerivative`: return a tuple containg the
+  Hosford equivalent stress, its first derivative (the normal) and the
+  second derivative.
+
+## Example
+
+The following example computes the Hosford equivalent stress, its
+normal and second derivative:
+
+~~~~{.cpp}
+stress seq;
+Stensor  n;
+Stensor4 dn;
+std::tie(seq,n,dn) = computeHosfordStressSecondDerivative(s,a,seps);
+~~~~
+
+In this example, `s` is the stress tensor, `a` is the Hosford
+exponent, `seps` is a numerical parameter used to detect when two
+eigenvalues are equal.
+
+If `C++-17` is available, the previous code can be made much more readable:
+
+~~~~{.cpp}
+const auto [seq,n,dn] = computeHosfordStressSecondDerivative(s,a,seps);
+~~~~
 
 # Release of `TFEL` `3.0.2` (25/10/2017)
 
@@ -797,7 +847,7 @@ of the `mfront-doc` functionalities can be found
 # New documentations pages (18 August 2015)
 
 New documentation pages were added to describe the `MTest` and
-`MFront` keywords~:
+`MFront` keywords:
 
 - `MFront` keywords, sorted by domain specific languages:
     + [DefaultDSL](DefaultDSL-keywords.html)